requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.10
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
idna==3.11
psycopg2-binary==2.9.11
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
